all messages working

+++ balance/src/main/resources/db/changelog/sql/balance_init.sql
@@ -2,7 +2,7 @@
 CREATE TABLE IF NOT EXISTS balances
 (
     id               SERIAL PRIMARY KEY,
-    account_id       INT        NOT NULL,
+    account_id       BIGINT     NOT NULL,
     currency         VARCHAR(3) NOT NULL,
     available_amount BIGINT DEFAULT 0.00,
     CONSTRAINT fk_account FOREIGN KEY (account_id) REFERENCES accounts (id)
